UF-4916 - Relative location header
Changes the location header to an absolute URI instead of a relative one
