+++ src/test/java/se/sundsvall/messaging/api/MessageResourceTests.java
@@ -7,6 +7,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.springframework.http.HttpStatus.CREATED;
+import static se.sundsvall.messaging.model.MessageStatus.FAILED;
 import static se.sundsvall.messaging.model.MessageStatus.SENT;
 import static se.sundsvall.messaging.model.MessageType.DIGITAL_MAIL;
 import static se.sundsvall.messaging.model.MessageType.EMAIL;
@@ -23,6 +24,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.web.util.UriComponentsBuilder;
 
 import se.sundsvall.messaging.api.model.request.DigitalMailRequest;
 import se.sundsvall.messaging.api.model.request.EmailRequest;
@@ -31,6 +33,7 @@ import se.sundsvall.messaging.api.model.request.MessageRequest;
 import se.sundsvall.messaging.api.model.request.SmsRequest;
 import se.sundsvall.messaging.api.model.request.SnailMailRequest;
 import se.sundsvall.messaging.api.model.request.WebMessageRequest;
+import se.sundsvall.messaging.api.model.response.DeliveryResult;
 import se.sundsvall.messaging.model.InternalDeliveryBatchResult;
 import se.sundsvall.messaging.model.InternalDeliveryResult;
 import se.sundsvall.messaging.service.MessageEventDispatcher;
@@ -47,6 +50,8 @@ class MessageResourceTests {
         .withStatus(SENT)
         .build();
 
+    private final UriComponentsBuilder uriComponentsBuilder = UriComponentsBuilder.newInstance();
+
     @Mock
     private MessageService mockMessageService;
     @Mock
@@ -60,7 +65,7 @@ class MessageResourceTests {
         when(mockMessageService.sendSms(any(SmsRequest.class)))
             .thenReturn(new InternalDeliveryResult("someMessageId", "someDeliveryId", SMS, SENT));
 
-        var response = messageResource.sendSms(SmsRequest.builder().build(), false);
+        var response = messageResource.sendSms(SmsRequest.builder().build(), false, uriComponentsBuilder);
 
         assertThat(response).isNotNull();
         assertThat(response.getStatusCode()).isEqualTo(CREATED);
@@ -80,7 +85,7 @@ class MessageResourceTests {
         when(mockEventDispatcher.handleSmsRequest(any(SmsRequest.class)))
             .thenReturn(deliveryResult.withMessageType(SMS));
 
-        var response = messageResource.sendSms(SmsRequest.builder().build(), true);
+        var response = messageResource.sendSms(SmsRequest.builder().build(), true, uriComponentsBuilder);
 
         assertThat(response).isNotNull();
         assertThat(response.getStatusCode()).isEqualTo(CREATED);
@@ -100,7 +105,7 @@ class MessageResourceTests {
         when(mockMessageService.sendEmail(any(EmailRequest.class)))
             .thenReturn(new InternalDeliveryResult("someMessageId", "someDeliveryId", EMAIL, SENT));
 
-        var response = messageResource.sendEmail(EmailRequest.builder().build(), false);
+        var response = messageResource.sendEmail(EmailRequest.builder().build(), false, uriComponentsBuilder);
 
         assertThat(response).isNotNull();
         assertThat(response.getStatusCode()).isEqualTo(CREATED);
@@ -120,7 +125,7 @@ class MessageResourceTests {
         when(mockEventDispatcher.handleEmailRequest(any(EmailRequest.class)))
             .thenReturn(deliveryResult.withMessageType(EMAIL));
 
-        var response = messageResource.sendEmail(EmailRequest.builder().build(), true);
+        var response = messageResource.sendEmail(EmailRequest.builder().build(), true, uriComponentsBuilder);
 
         assertThat(response).isNotNull();
         assertThat(response.getStatusCode()).isEqualTo(CREATED);
@@ -140,7 +145,7 @@ class MessageResourceTests {
         when(mockMessageService.sendWebMessage(any(WebMessageRequest.class)))
             .thenReturn(new InternalDeliveryResult("someMessageId", "someDeliveryId", WEB_MESSAGE, SENT));
 
-        var response = messageResource.sendWebMessage(WebMessageRequest.builder().build(), false);
+        var response = messageResource.sendWebMessage(WebMessageRequest.builder().build(), false, uriComponentsBuilder);
 
         assertThat(response).isNotNull();
         assertThat(response.getStatusCode()).isEqualTo(CREATED);
@@ -160,7 +165,7 @@ class MessageResourceTests {
         when(mockEventDispatcher.handleWebMessageRequest(any(WebMessageRequest.class)))
             .thenReturn(deliveryResult.withMessageType(WEB_MESSAGE));
 
-        var response = messageResource.sendWebMessage(WebMessageRequest.builder().build(), true);
+        var response = messageResource.sendWebMessage(WebMessageRequest.builder().build(), true, uriComponentsBuilder);
 
         assertThat(response).isNotNull();
         assertThat(response.getStatusCode()).isEqualTo(CREATED);
@@ -181,7 +186,7 @@ class MessageResourceTests {
             .thenReturn(new InternalDeliveryBatchResult("someBatchId",
                 List.of(new InternalDeliveryResult("someMessageId", "someDeliveryId", DIGITAL_MAIL, SENT))));
 
-        var response = messageResource.sendDigitalMail(DigitalMailRequest.builder().build(), false);
+        var response = messageResource.sendDigitalMail(DigitalMailRequest.builder().build(), false, uriComponentsBuilder);
 
         assertThat(response).isNotNull();
         assertThat(response.getStatusCode()).isEqualTo(CREATED);
@@ -210,7 +215,7 @@ class MessageResourceTests {
                 .withDeliveries(List.of(deliveryResult.withMessageType(DIGITAL_MAIL)))
                 .build());
 
-        var response = messageResource.sendDigitalMail(DigitalMailRequest.builder().build(), true);
+        var response = messageResource.sendDigitalMail(DigitalMailRequest.builder().build(), true, uriComponentsBuilder);
 
         assertThat(response).isNotNull();
         assertThat(response.getStatusCode()).isEqualTo(CREATED);
@@ -236,7 +241,7 @@ class MessageResourceTests {
             .thenReturn(new InternalDeliveryBatchResult("someBatchId",
                 List.of(new InternalDeliveryResult("someMessageId", "someDeliveryId", MESSAGE, SENT))));
 
-        var response = messageResource.sendMessages(MessageRequest.builder().build(), false);
+        var response = messageResource.sendMessages(MessageRequest.builder().build(), false, uriComponentsBuilder);
 
         assertThat(response).isNotNull();
         assertThat(response.getStatusCode()).isEqualTo(CREATED);
@@ -264,7 +269,7 @@ class MessageResourceTests {
                 .withDeliveries(List.of(deliveryResult.withMessageType(MESSAGE)))
                 .build());
 
-        var response = messageResource.sendMessages(MessageRequest.builder().build(), true);
+        var response = messageResource.sendMessages(MessageRequest.builder().build(), true, uriComponentsBuilder);
 
         assertThat(response).isNotNull();
         assertThat(response.getStatusCode()).isEqualTo(CREATED);
@@ -289,7 +294,7 @@ class MessageResourceTests {
         when(mockMessageService.sendSnailMail(any(SnailMailRequest.class)))
             .thenReturn(new InternalDeliveryResult("someMessageId", "someDeliveryId", SNAIL_MAIL, SENT));
 
-        var response = messageResource.sendSnailMail(SnailMailRequest.builder().build(), false);
+        var response = messageResource.sendSnailMail(SnailMailRequest.builder().build(), false, uriComponentsBuilder);
 
         assertThat(response).isNotNull();
         assertThat(response.getStatusCode()).isEqualTo(CREATED);
@@ -309,7 +314,7 @@ class MessageResourceTests {
         when(mockEventDispatcher.handleSnailMailRequest(any(SnailMailRequest.class)))
             .thenReturn(deliveryResult.withMessageType(SNAIL_MAIL));
 
-        var response = messageResource.sendSnailMail(SnailMailRequest.builder().build(), true);
+        var response = messageResource.sendSnailMail(SnailMailRequest.builder().build(), true, uriComponentsBuilder);
 
         assertThat(response).isNotNull();
         assertThat(response.getStatusCode()).isEqualTo(CREATED);
@@ -330,7 +335,7 @@ class MessageResourceTests {
             .thenReturn(new InternalDeliveryBatchResult("someBatchId",
                 List.of(new InternalDeliveryResult("someMessageId", "someDeliveryId", LETTER, SENT))));
 
-        var response = messageResource.sendLetter(LetterRequest.builder().build(), false);
+        var response = messageResource.sendLetter(LetterRequest.builder().build(), false, uriComponentsBuilder);
 
         assertThat(response).isNotNull();
         assertThat(response.getStatusCode()).isEqualTo(CREATED);
@@ -358,7 +363,7 @@ class MessageResourceTests {
                 .withDeliveries(List.of(deliveryResult.withMessageType(LETTER)))
                 .build());
 
-        var response = messageResource.sendLetter(LetterRequest.builder().build(), true);
+        var response = messageResource.sendLetter(LetterRequest.builder().build(), true, uriComponentsBuilder);
 
         assertThat(response).isNotNull();
         assertThat(response.getStatusCode()).isEqualTo(CREATED);
@@ -376,5 +381,66 @@ class MessageResourceTests {
 
         verify(mockMessageService, never()).sendLetter(any(LetterRequest.class));
         verify(mockEventDispatcher, times(1)).handleLetterRequest(any(LetterRequest.class));
+    }
+
+    @Test
+    void test_toResponse_fromDeliveryResult() {
+        var deliveryResult = InternalDeliveryResult.builder()
+            .withMessageId("someMessageId")
+            .withDeliveryId("someDeliveryId")
+            .withMessageType(DIGITAL_MAIL)
+            .withStatus(SENT)
+            .build();
+
+        var result = messageResource.toResponse(uriComponentsBuilder, deliveryResult);
+
+        assertThat(result).isNotNull();
+        assertThat(result.getStatusCode()).isEqualTo(CREATED);
+        assertThat(result.getBody()).isNotNull().satisfies(message -> {
+            assertThat(message.messageId()).isEqualTo(deliveryResult.messageId());
+            assertThat(message.deliveries()).hasSize(1).allSatisfy(delivery -> {
+                assertThat(delivery.deliveryId()).isEqualTo(deliveryResult.deliveryId());
+                assertThat(delivery.messageType()).isEqualTo(deliveryResult.messageType());
+                assertThat(delivery.status()).isEqualTo(deliveryResult.status());
+            });
+        });
+    }
+
+    @Test
+    void test_toResponse_fromBatchResult() {
+        var deliveryBatchResult = InternalDeliveryBatchResult.builder()
+            .withBatchId("someBatchId")
+            .withDeliveries(List.of(
+                InternalDeliveryResult.builder()
+                    .withMessageId("someMessageId")
+                    .withDeliveryId("someDeliveryId")
+                    .withMessageType(DIGITAL_MAIL)
+                    .withStatus(FAILED)
+                    .build(),
+                InternalDeliveryResult.builder()
+                    .withMessageId("someMessageId")
+                    .withDeliveryId("someOtherDeliveryId")
+                    .withMessageType(SNAIL_MAIL)
+                    .withStatus(SENT)
+                    .build()
+            ))
+            .build();
+
+        var result = messageResource.toResponse(uriComponentsBuilder, deliveryBatchResult);
+
+        assertThat(result).isNotNull();
+        assertThat(result.getStatusCode()).isEqualTo(CREATED);
+        assertThat(result.getBody()).isNotNull().satisfies(batch -> {
+            assertThat(batch.batchId()).isEqualTo(deliveryBatchResult.batchId());
+            assertThat(batch.messages()).hasSize(1).allSatisfy(message -> {
+                assertThat(message.messageId()).isNotNull();
+                assertThat(message.deliveries()).hasSize(2);
+                assertThat(message.deliveries()).extracting(DeliveryResult::deliveryId).doesNotContainNull();
+                assertThat(message.deliveries()).extracting(DeliveryResult::messageType)
+                    .containsExactlyInAnyOrder(DIGITAL_MAIL, SNAIL_MAIL);
+                assertThat(message.deliveries()).extracting(DeliveryResult::status)
+                    .containsExactlyInAnyOrder(FAILED, SENT);
+            });
+        });
     }
 }
